+++ unit_tests/reduct/storage/entry_test.cc
@@ -13,6 +13,7 @@ using reduct::ReadOne;
 using reduct::WriteOne;
 using reduct::core::Error;
 using reduct::core::Time;
+using reduct::core::ToMicroseconds;
 using reduct::storage::IEntry;
 
 using google::protobuf::util::TimeUtil;
@@ -31,15 +32,9 @@ static auto MakeDefaultOptions() {
 
 static const auto kTimestamp = Time() + std::chrono::microseconds(10'100'200);  // to check um precision
 
-auto ToMicroseconds(Time tp) {
-  return std::chrono::duration_cast<std::chrono::microseconds>(tp.time_since_epoch()).count();
-}
 
 auto GetBlockSize(std::string_view name, fs::path path, IEntry::Options options, Time begin_ts) {
-  return fs::file_size(
-      path / name /
-      fmt::format("{}.blk",
-                  std::chrono::duration_cast<std::chrono::microseconds>(begin_ts.time_since_epoch()).count()));
+  return fs::file_size(path / name / fmt::format("{}.blk", ToMicroseconds(begin_ts)));
 }
 
 TEST_CASE("storage::Entry should record data to a block", "[entry]") {
@@ -165,6 +160,22 @@ TEST_CASE("storage::Entry should create block with size of record it  is bigger 
   REQUIRE(WriteOne(*entry, big_data, kTimestamp) == Error::kOk);
 
   REQUIRE(GetBlockSize(kName, path, options, kTimestamp) == big_data.size());
+}
+
+TEST_CASE("storage::Entry should create block if previous invalid", "[entry][block]") {
+  const auto options = MakeDefaultOptions();
+  const auto path = BuildTmpDirectory();
+  auto entry = IEntry::Build(kName, path, options);
+  REQUIRE(entry);
+
+  std::string data = "xxxx";
+  REQUIRE(WriteOne(*entry, data, kTimestamp) == Error::kOk);
+  // remove block with data to make write operation invalid
+  REQUIRE(fs::remove(path / kName / fmt::format("{}.blk", ToMicroseconds(kTimestamp))));
+  REQUIRE(WriteOne(*entry, data, kTimestamp + seconds(1)) == Error{.code = 500, .message = "Bad block"});
+
+  REQUIRE(WriteOne(*entry, data, kTimestamp + seconds(2)) == Error::kOk);
+  REQUIRE(GetBlockSize(kName, path, options, kTimestamp + seconds(2)) == 100);
 }
 
 TEST_CASE("storage::Entry should write data for random kTimestamp", "[entry]") {