+++ CHANGELOG.md
@@ -7,15 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## Fixed:
+### Fixed:
 
 - Infinite loop in Bucket::KeepQuota, [PR-146](https://github.com/reduct-storage/reduct-storage/pull/146)
 - waiting data from HTTP client if it aborts
   connection, [PR-151](https://github.com/reduct-storage/reduct-storage/pull/151)
 
-## Changed:
+### Changed:
 
-- Duplication of timestamps is not allowed, [PR147](https://github.com/reduct-storage/reduct-storage/pull/147)
+- Duplication of timestamps is not allowed, [PR-147](https://github.com/reduct-storage/reduct-storage/pull/147)
+
+### Fixed:
+
+- writing record when current block is broken, [PR-15-](https://github.com/reduct-storage/reduct-storage/pull/153)
 
 ## [0.7.1] - 2022-07-30